Fix data race when threads raise an error msg at the same time

See merge request gmsh/gmsh!466

+++ utils/solvers/c++/GmshSocket.h
@@ -28,6 +28,7 @@
 //#include "GmshConfig.h"
 
 #include <string>
+#include <stdexcept>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -55,42 +56,44 @@ typedef int socklen_t;
 typedef int socklen_t;
 #endif
 
-class GmshSocket{
- public:
+class GmshSocket {
+public:
   // types of messages that can be exchanged (never use values greater
   // that 65535: if we receive a type > 65535 we assume that we
   // receive data from a machine with a different byte ordering, and
   // we swap the bytes in the payload)
-  enum MessageType{
-    GMSH_START               = 1,
-    GMSH_STOP                = 2,
-    GMSH_INFO                = 10,
-    GMSH_WARNING             = 11,
-    GMSH_ERROR               = 12,
-    GMSH_PROGRESS            = 13,
-    GMSH_MERGE_FILE          = 20,
-    GMSH_PARSE_STRING        = 21,
-    GMSH_VERTEX_ARRAY        = 22,
-    GMSH_PARAMETER           = 23,
-    GMSH_PARAMETER_QUERY     = 24,
+  enum MessageType {
+    GMSH_START = 1,
+    GMSH_STOP = 2,
+    GMSH_INFO = 10,
+    GMSH_WARNING = 11,
+    GMSH_ERROR = 12,
+    GMSH_PROGRESS = 13,
+    GMSH_MERGE_FILE = 20,
+    GMSH_PARSE_STRING = 21,
+    GMSH_VERTEX_ARRAY = 22,
+    GMSH_PARAMETER = 23,
+    GMSH_PARAMETER_QUERY = 24,
     GMSH_PARAMETER_QUERY_ALL = 25,
     GMSH_PARAMETER_QUERY_END = 26,
-    GMSH_CONNECT             = 27,
-    GMSH_OLPARSE             = 28,
+    GMSH_CONNECT = 27,
+    GMSH_OLPARSE = 28,
     GMSH_PARAMETER_NOT_FOUND = 29,
-    GMSH_SPEED_TEST          = 30,
-    GMSH_PARAMETER_CLEAR     = 31,
-    GMSH_PARAMETER_UPDATE    = 32,
-    GMSH_OPEN_PROJECT        = 33,
-    GMSH_CLIENT_CHANGED      = 34,
+    GMSH_SPEED_TEST = 30,
+    GMSH_PARAMETER_CLEAR = 31,
+    GMSH_PARAMETER_UPDATE = 32,
+    GMSH_OPEN_PROJECT = 33,
+    GMSH_CLIENT_CHANGED = 34,
     GMSH_PARAMETER_WITHOUT_CHOICES = 35,
     GMSH_PARAMETER_QUERY_WITHOUT_CHOICES = 36,
-    GMSH_OPTION_1            = 100,
-    GMSH_OPTION_2            = 101,
-    GMSH_OPTION_3            = 102,
-    GMSH_OPTION_4            = 103,
-    GMSH_OPTION_5            = 104};
- protected:
+    GMSH_OPTION_1 = 100,
+    GMSH_OPTION_2 = 101,
+    GMSH_OPTION_3 = 102,
+    GMSH_OPTION_4 = 103,
+    GMSH_OPTION_5 = 104
+  };
+
+protected:
   // the socket descriptor
   int _sock;
   // the socket name
@@ -135,10 +138,9 @@ class GmshSocket{
     for(int i = 0; i < n; i++) {
       char *a = &array[i * size];
       memcpy(x, a, size);
-      for(int c = 0; c < size; c++)
-        a[size - 1 - c] = x[c];
+      for(int c = 0; c < size; c++) a[size - 1 - c] = x[c];
     }
-    delete [] x;
+    delete[] x;
   }
   // sleep for some milliseconds
   void _sleep(int ms)
@@ -149,7 +151,8 @@ class GmshSocket{
     Sleep(ms);
 #endif
   }
- public:
+
+public:
   GmshSocket() : _sock(0), _sent(0), _received(0)
   {
 #if defined(WIN32) && !defined(__CYGWIN__)
@@ -167,7 +170,7 @@ class GmshSocket{
   // we check for available data and return immediately, i.e., we do
   // polling). Returns 1 when data is available, 0 when nothing happened before
   // the time delay, -1 on error.
-  int Select(int seconds, int microseconds, int socket=-1)
+  int Select(int seconds, int microseconds, int socket = -1)
   {
     int s = (socket < 0) ? _sock : socket;
     struct timeval tv;
@@ -192,14 +195,14 @@ class GmshSocket{
   {
     SendMessage(type, (int)strlen(str), str);
   }
-  void Info(const char *str){ SendString(GMSH_INFO, str); }
-  void Warning(const char *str){ SendString(GMSH_WARNING, str); }
-  void Error(const char *str){ SendString(GMSH_ERROR, str); }
-  void Progress(const char *str){ SendString(GMSH_PROGRESS, str); }
-  void MergeFile(const char *str){ SendString(GMSH_MERGE_FILE, str); }
-  void OpenProject(const char *str){ SendString(GMSH_OPEN_PROJECT, str); }
-  void ParseString(const char *str){ SendString(GMSH_PARSE_STRING, str); }
-  void SpeedTest(const char *str){ SendString(GMSH_SPEED_TEST, str); }
+  void Info(const char *str) { SendString(GMSH_INFO, str); }
+  void Warning(const char *str) { SendString(GMSH_WARNING, str); }
+  void Error(const char *str) { SendString(GMSH_ERROR, str); }
+  void Progress(const char *str) { SendString(GMSH_PROGRESS, str); }
+  void MergeFile(const char *str) { SendString(GMSH_MERGE_FILE, str); }
+  void OpenProject(const char *str) { SendString(GMSH_OPEN_PROJECT, str); }
+  void ParseString(const char *str) { SendString(GMSH_PARSE_STRING, str); }
+  void SpeedTest(const char *str) { SendString(GMSH_SPEED_TEST, str); }
   void Option(int num, const char *str)
   {
     if(num < 1) num = 1;
@@ -209,15 +212,15 @@ class GmshSocket{
   int ReceiveHeader(int *type, int *len, int *swap)
   {
     *swap = 0;
-    if(_receiveData(type, sizeof(int)) > 0){
-      if(*type > 65535){
+    if(_receiveData(type, sizeof(int)) > 0) {
+      if(*type > 65535) {
         // the data comes from a machine with different endianness and
         // we must swap the bytes
         *swap = 1;
-        _swapBytes((char*)type, sizeof(int), 1);
+        _swapBytes((char *)type, sizeof(int), 1);
       }
-      if(_receiveData(len, sizeof(int)) > 0){
-        if(*swap) _swapBytes((char*)len, sizeof(int), 1);
+      if(_receiveData(len, sizeof(int)) > 0) {
+        if(*swap) _swapBytes((char *)len, sizeof(int), 1);
         return 1;
       }
     }
@@ -251,24 +254,25 @@ class GmshSocket{
     shutdown(s, SHUT_RDWR);
 #endif
   }
-  unsigned long int SentBytes(){ return _sent; }
-  unsigned long int ReceivedBytes(){ return _received; }
+  unsigned long int SentBytes() { return _sent; }
+  unsigned long int ReceivedBytes() { return _received; }
 };
 
 class GmshClient : public GmshSocket {
- public:
+public:
   GmshClient() : GmshSocket() {}
-  ~GmshClient(){}
+  ~GmshClient() {}
   int Connect(const char *sockname)
   {
-    if(strstr(sockname, "/") || strstr(sockname, "\\") || !strstr(sockname, ":")){
+    if(strstr(sockname, "/") || strstr(sockname, "\\") ||
+       !strstr(sockname, ":")) {
 #if !defined(WIN32) || defined(__CYGWIN__)
       // UNIX socket (testing ":" is not enough with Windows paths)
       _sock = socket(PF_UNIX, SOCK_STREAM, 0);
       if(_sock < 0) return -1;
       // try to connect socket to given name
       struct sockaddr_un addr_un;
-      memset((char *) &addr_un, 0, sizeof(addr_un));
+      memset((char *)&addr_un, 0, sizeof(addr_un));
       addr_un.sun_family = AF_UNIX;
       strcpy(addr_un.sun_path, sockname);
       for(int tries = 0; tries < 5; tries++) {
@@ -280,7 +284,7 @@ class GmshClient : public GmshSocket {
       return -1; // Unix sockets are not available on Windows
 #endif
     }
-    else{
+    else {
       // TCP/IP socket
       _sock = socket(AF_INET, SOCK_STREAM, 0);
       if(_sock < 0) return -1;
@@ -292,26 +296,25 @@ class GmshClient : public GmshSocket {
       int portno = atoi(port + 1);
       char *remote = strdup(sockname);
       int remotelen = (int)(strlen(remote) - strlen(port));
-      if(remotelen > 0)
-        strncpy(remote, sockname, remotelen);
-      if(remotelen >= 0)
-        remote[remotelen] = '\0';
+      if(remotelen > 0) strncpy(remote, sockname, remotelen);
+      if(remotelen >= 0) remote[remotelen] = '\0';
       struct hostent *server;
-      if(!(server = gethostbyname(remote))){
+      if(!(server = gethostbyname(remote))) {
         CloseSocket(_sock);
         free(remote);
         return -3; // no such host
       }
       free(remote);
       struct sockaddr_in addr_in;
-      memset((char *) &addr_in, 0, sizeof(addr_in));
+      memset((char *)&addr_in, 0, sizeof(addr_in));
       addr_in.sin_family = AF_INET;
-      memcpy((char *)&addr_in.sin_addr.s_addr, (char *)server->h_addr, server->h_length);
+      memcpy((char *)&addr_in.sin_addr.s_addr, (char *)server->h_addr,
+             server->h_length);
       addr_in.sin_port = htons(portno);
       for(int tries = 0; tries < 5; tries++) {
-        if(connect(_sock, (struct sockaddr *)&addr_in, sizeof(addr_in)) >= 0){
+        if(connect(_sock, (struct sockaddr *)&addr_in, sizeof(addr_in)) >= 0) {
           return _sock;
-	}
+        }
         _sleep(100);
       }
     }
@@ -328,27 +331,30 @@ class GmshClient : public GmshSocket {
 #endif
     SendString(GMSH_START, tmp);
   }
-  void Stop(){ SendString(GMSH_STOP, "Goodbye!"); }
-  void Disconnect(){ CloseSocket(_sock); }
+  void Stop() { SendString(GMSH_STOP, "Goodbye!"); }
+  void Disconnect() { CloseSocket(_sock); }
 };
 
-class GmshServer : public GmshSocket{
- private:
+class GmshServer : public GmshSocket {
+private:
   int _portno;
- public:
+
+public:
   GmshServer() : GmshSocket(), _portno(-1) {}
-  virtual ~GmshServer(){}
-  virtual int NonBlockingSystemCall(const std::string &exe, const std::string &args) = 0;
-  virtual int NonBlockingWait(double waitint, double timeout, int socket=-1) = 0;
+  virtual ~GmshServer() {}
+  virtual int NonBlockingSystemCall(const std::string &exe,
+                                    const std::string &args) = 0;
+  virtual int NonBlockingWait(double waitint, double timeout,
+                              int socket = -1) = 0;
   // start the client by launching "exe args" (args is supposed to contain
   // '%s' where the socket name should appear)
-  int Start(const std::string &exe, const std::string &args, const std::string &sockname,
-            double timeout)
+  int Start(const std::string &exe, const std::string &args,
+            const std::string &sockname, double timeout)
   {
     _sockname = sockname;
     int tmpsock;
     if(strstr(_sockname.c_str(), "/") || strstr(_sockname.c_str(), "\\") ||
-       !strstr(_sockname.c_str(), ":")){
+       !strstr(_sockname.c_str(), ":")) {
       // UNIX socket (testing ":" is not enough with Windows paths)
       _portno = -1;
 #if !defined(WIN32) || defined(__CYGWIN__)
@@ -356,23 +362,23 @@ class GmshServer : public GmshSocket{
       unlink(_sockname.c_str());
       // create a socket
       tmpsock = socket(PF_UNIX, SOCK_STREAM, 0);
-      if(tmpsock < 0) throw "Couldn't create socket";
+      if(tmpsock < 0) throw std::runtime_error("Couldn't create socket");
       // bind the socket to its name
       struct sockaddr_un addr_un;
-      memset((char *) &addr_un, 0, sizeof(addr_un));
+      memset((char *)&addr_un, 0, sizeof(addr_un));
       strcpy(addr_un.sun_path, _sockname.c_str());
       addr_un.sun_family = AF_UNIX;
-      if(bind(tmpsock, (struct sockaddr *)&addr_un, sizeof(addr_un)) < 0){
+      if(bind(tmpsock, (struct sockaddr *)&addr_un, sizeof(addr_un)) < 0) {
         CloseSocket(tmpsock);
-        throw "Couldn't bind socket to name";
+        throw std::runtime_error("Couldn't bind socket to name");
       }
       // change permissions on the socket name in case it has to be rm'd later
       chmod(_sockname.c_str(), 0666);
 #else
-      throw "Unix sockets not available on Windows";
+      throw std::runtime_error("Unix sockets not available on Windows");
 #endif
     }
-    else{
+    else {
       // TCP/IP socket: valid names are either explicit ("hostname:12345")
       // or implicit ("hostname:", "hostname: ", "hostname:0") in which case
       // the system attributes at random an available port
@@ -388,65 +394,63 @@ class GmshServer : public GmshSocket{
 #else
       if(tmpsock == (int)INVALID_SOCKET)
 #endif
-        throw "Couldn't create socket";
+        throw std::runtime_error("Couldn't create socket");
       // bind the socket to its name
       struct sockaddr_in addr_in;
-      memset((char *) &addr_in, 0, sizeof(addr_in));
+      memset((char *)&addr_in, 0, sizeof(addr_in));
       addr_in.sin_family = AF_INET;
       addr_in.sin_addr.s_addr = INADDR_ANY;
       addr_in.sin_port = htons(_portno); // random assign if _portno == 0
-      if(bind(tmpsock, (struct sockaddr *)&addr_in, sizeof(addr_in)) < 0){
+      if(bind(tmpsock, (struct sockaddr *)&addr_in, sizeof(addr_in)) < 0) {
         CloseSocket(tmpsock);
-        throw "Couldn't bind socket to name";
+        throw std::runtime_error("Couldn't bind socket to name");
       }
-      if(!_portno){ // retrieve name if randomly assigned port
+      if(!_portno) { // retrieve name if randomly assigned port
         socklen_t addrlen = sizeof(addr_in);
         getsockname(tmpsock, (struct sockaddr *)&addr_in, &addrlen);
         _portno = ntohs(addr_in.sin_port);
-	int pos = (int)_sockname.find(':'); // remove trailing ' ' or '0'
+        int pos = (int)_sockname.find(':'); // remove trailing ' ' or '0'
         char tmp[256];
-	sprintf(tmp, "%s:%d", _sockname.substr(0, pos).c_str(), _portno);
+        sprintf(tmp, "%s:%d", _sockname.substr(0, pos).c_str(), _portno);
         _sockname.assign(tmp);
       }
     }
 
-    if(exe.size() || args.size()){
+    if(exe.size() || args.size()) {
       char s[1024];
       sprintf(s, args.c_str(), _sockname.c_str());
       NonBlockingSystemCall(exe, s); // starts the solver
     }
-    else{
+    else {
       timeout = 0.; // no command launched: don't set a timeout
     }
 
     // listen on socket (queue up to 20 connections before having
     // them automatically rejected)
-    if(listen(tmpsock, 20)){
+    if(listen(tmpsock, 20)) {
       CloseSocket(tmpsock);
-      throw "Socket listen failed";
+      throw std::runtime_error("Socket listen failed");
     }
 
     // wait until we get data
     int ret = NonBlockingWait(0.001, timeout, tmpsock);
-    if(ret){
+    if(ret) {
       CloseSocket(tmpsock);
-      if(ret == 2){
-        throw "Socket listening timeout";
-      }
-      else{
+      if(ret == 2) { throw std::runtime_error("Socket listening timeout"); }
+      else {
         return -1; // stopped listening
       }
     }
 
     // accept connection request
-    if(_portno < 0){
+    if(_portno < 0) {
 #if !defined(WIN32) || defined(__CYGWIN__)
       struct sockaddr_un from_un;
       socklen_t len = sizeof(from_un);
       _sock = accept(tmpsock, (struct sockaddr *)&from_un, &len);
 #endif
     }
-    else{
+    else {
       struct sockaddr_in from_in;
       socklen_t len = sizeof(from_in);
       _sock = accept(tmpsock, (struct sockaddr *)&from_in, &len);
@@ -454,15 +458,13 @@ class GmshServer : public GmshSocket{
       setsockopt(_sock, IPPROTO_TCP, TCP_NODELAY, &one, sizeof(one));
     }
     CloseSocket(tmpsock);
-    if(_sock < 0)
-      throw "Socket accept failed";
+    if(_sock < 0) throw std::runtime_error("Socket accept failed");
     return _sock;
   }
   int Shutdown()
   {
 #if !defined(WIN32) || defined(__CYGWIN__)
-    if(_portno < 0)
-      unlink(_sockname.c_str());
+    if(_portno < 0) unlink(_sockname.c_str());
 #endif
     ShutdownSocket(_sock);
     CloseSocket(_sock);
